+++ simulation/simple_sensor_simulator/CMakeLists.txt
@@ -35,6 +35,7 @@ ament_auto_add_library(simple_sensor_simulator_component SHARED
   src/sensor_simulation/lidar/lidar_sensor.cpp
   src/sensor_simulation/lidar/raycaster.cpp
   src/sensor_simulation/occupancy_grid/grid.cpp
+  src/sensor_simulation/occupancy_grid/grid_cell.cpp
   src/sensor_simulation/occupancy_grid/occupancy_grid_generator.cpp
   src/sensor_simulation/occupancy_grid/occupancy_grid_sensor.cpp
   src/sensor_simulation/primitives/box.cpp
