+++ simulation/simple_sensor_simulator/src/sensor_simulation/occupancy_grid/occupancy_grid_generator.cpp
@@ -40,15 +40,9 @@ nav_msgs::msg::OccupancyGrid OccupancyGridGenerator::generate(
   occupancy_grid.info.origin.position.y = occupancy_grid.info.origin.position.y -
                                           0.5 * configuration.width() * configuration.resolution();
   for (const auto & primitive : primitive_ptrs_) {
-    for (const auto & cell : grid.getInvisibleCell(primitive.second)) {
-      occupancy_grid.data[cell.index] = 50;
-    }
-    /*
-    for (const auto & cell : grid.getOccupiedCell(primitive.second)) {
-      occupancy_grid.data[cell.index] = 100;
-    }
-    */
+    grid.addPrimitive(primitive.second);
   }
+  occupancy_grid.data = grid.getData();
   return occupancy_grid;
 }
 }  // namespace simple_sensor_simulator
